feat: start implementing cleartext dialing (#2)
We're not quite there yet in terms of emitting events, but we have made
some progress towards what we actually need.

Loosely base the code on OONI Probe code for now, by cherry picking and
going on memory while writing this.

Also, document my current doubts regarding whether it would be better to
have a unified type for dialing all connections types or whether
separate types would be better. For now, keep the separate types and
re-evaluate this in the future.