refactor(netcore): Dialer+TLSDialer = Network
This commit takes stock of the fact that having two separate types
would complicate sharing code and that, in 2024, TLS is so integral
to networking that it *is* part of the basic networking.

As such, we just refactor Dialer and TLSDialer into a single type
called Network and name DialTLSContext the TLS dialing func.