fix(netstack): repair reply from listening UDP socket
To this end, we need to handle the case where we're sending and
the source address is unspecified, by finding the proper source
address for the destination address.

While there, significantly fix how we determine the proper
source address by precisely mapping the address family, and
by adding support for IPv4-mapped IPv6 addresses.