feat: add packet filtering support (#26)
This commit adds support for filtering packets in the router.

The interfaces for filtering packets go into `packet/packet.go` since
they are integral to our network simulation.

We implementing filtering in the router because we need the ability to
inject arbitrary packets in all directions. It won't be possible to
implement filtering with links, since they are unidirectional (i.e., one
direction of a given link does not have access to the link direction).