BREAKING CHANGE: {conn => close}pool and remove New() (#30)
This diff renames the connpool package to closepool, since we reckon
that is the correct naming. A connpool would also allow to get/put
connections, which the package does not allow for.

Hence, choose a more appropriate naming.

While there, drop the New constructor that was just constructing an
empty value already ready to use.