fix(netsim): forward return traffic via geolink
The way in which we were adding a geolink to a stack was such
that the return route was not routed through the geolink.

This was obvious when trying to print packets.

To fix, rework how we attach a network device to a router to
combine reading from the device and setting up the return route,
also make sure we propagate the IP addresses of the device to
the other geolink device, so the router can setup routes.