BREAKING CHANGE: netcore: scope non-common fields
1. `domain` => `dnsLookupDomain`

2. `addrs` => `dnsResolvedAddrs`

3. `count` => `ioBufferSize` or `ioBufferCount`

This set of changes is part of an overall effort to ensure it's
relatively trivial to serialize to a flat data format.

To this end, figure out which fields are really common and which
fields, instead, are specific, and add prefixes to the specific ones
to make it very clear what they are about.

Also, avoid overload, and split into two fields when we're using
the same field with two distinct semantics dependent on the
message type. This should also increase the data format orthogonality.

See also https://github.com/rbmk-project/dnscore/pull/13