feat(netcore): implement and use error classification (#48)
This diff implements and uses error classification to reduce errors to a
narrow enumeration set.

We keep the original error into the structured log, which helps with
debugging what happened in case we cannot map the error that occurred
with an enumerated error.

We're not using OONI errors because they are a bit weird and have some
legacy. In any case, we will be able to map to OONI errors, shall we
ever decide to reshaphe the results and submit them as OONI
measurements.