feat(netcore): start covering conn.go code
This diff starts adding tests to increase the coverage of code
in the netcore package and specifically in conn.go.

We want to have reasonably good coverage of functionality and error
paths for the netcore package ahead of migrating this code, which
now seems to be stable, to its own package.

While there, upgrade dependencies to silence an otherwise harmless
security advisory (we do not use `x/net/html` here).