fix(netcore): make connWrapper robust to nil Network.Logger (#57)
This diff fixes a bug where one could create a `*netcore.connWrapper`
using `netcore.WrapConn` with a `*netcore.Network` containing a `nil`
`Logger` field. The resulting connection would crash when attempting to
log.

Discovered while adding unit tests for the `connWrapper.Close`, which
are also included in this pull request.