fix(netcore): allow for utls-aware structured logs (#73)
This diff modifies `Network` in a backwards compatible way to add a new
optionally-nil field named `TLSEngine`.

This field is now the preferred way for creating a new `TLSConn` while
the previous `NewTLSClientConn` field, which is still supported, has now
become deprecated.

The problem we want to solve is that of making sure the name of the TLS
engine and the parrot being used are inside of the structured logs.

With `TLSEngine`, which is an interface, we know that, via the `Name()`
and `Parrot()` methods.

With `NewTLSClientConn`, we don't know, hence we set both to
`"unknown"`.